+++ RayTracerChallenge/Color.cpp
@@ -1,7 +1,7 @@
 #include "Color.h"
 #include "Helper.h"
 
-Color::Color(float x, float y, float z) : r(x), g(y), b(z) {};
+Color::Color(double x, double y, double z) : r(x), g(y), b(z) {};
 
 Color operator+(const Color& firstColor, const Color& secondColor) {
 	return Color(firstColor.r + secondColor.r, firstColor.g + secondColor.g, firstColor.b + secondColor.b);
@@ -11,7 +11,7 @@ Color operator-(const Color& firstColor, const Color& secondColor) {
 	return Color(firstColor.r - secondColor.r, firstColor.g - secondColor.g, firstColor.b - secondColor.b);
 }
 
-Color operator*(const Color& c, const float scalar) {
+Color operator*(const Color& c, const double scalar) {
 	return Color(c.r * scalar, c.g * scalar, c.b * scalar);
 }
