+++ RayTracerChallenge/Color.h
@@ -3,12 +3,12 @@
 
 struct Color {
 public:
-	Color(float r, float g, float b);
-	float r, g, b;
+	Color(double r, double g, double b);
+	double r, g, b;
 };
 
 bool operator==(const Color& firstColor, const Color& secondColor);
 Color operator+(const Color& firstColor, const Color& secondColor);
 Color operator-(const Color& firstColor, const Color& secondColor);
-Color operator*(const Color& color, float scalar);
+Color operator*(const Color& color, double scalar);
 Color operator*(const Color& firstColor, const Color& secondColor);
