+++ RayTracerChallenge/Material.h
@@ -8,10 +8,10 @@ struct Material {
 	Material():color(Color(1, 1, 1)) {}
 
 	Color color;
-	float ambient = 0.1;
-	float diffuse = 0.9;
-	float specular = 0.9;
-	float shininess = 200;
+	double ambient = 0.1;
+	double diffuse = 0.9;
+	double specular = 0.9;
+	double shininess = 200;
 
 	Color lighting(const Light& light, const Tuple& point, const Tuple& eyev, const Tuple& normalv, const bool inShadow)const;
 };
